requirements.txt
Flask
pika
PyMySQL
pymemcache
python-dotenv